requirements.txt
openai==1.30.1
streamlit==1.34.0
python-dotenv==1.0.1